imbalanced-exp/requirements.txt
matplotlib>=3.5.0
seaborn>=0.11.0
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=0.11.0
pyyaml>=6.0
numpy>=1.21.0
datasets>=2.0.0
huggingface_hub>=0.16.0
Pillow>=9.0.0
